Querying for control input by participant
This change adds the ability to query for control state by participant.

Ability to disable, and set progress on controls.

Note: the change appears large because it renames the "External" folder to "Tools". This was done to make easier to build on the components which looks for a folder named Tools to find it's dependencies.