feat(batch): Initial commit for batch operations
Exposing a new batch operations API that allows you to batch updates to
multiple objects (e.g. controls, participants) and multiple property
updates on a single object. This API also allows for the use of custom
data including nested objects and arrays.

Current only exposes batch methods for `controls` - other objects will
come in a subsequent commit.

Tests are not yet complete as we have yet to expose handlers for control
update events and the current implementation does not update internal
state on all events so it is not testable. We also need to invest in
proper teardown methods to kill interactive connection when using
asserts.